requirements.txt
whisperx
feedparser
python-dotenv
requests
torch
